more progress
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/branches/lucene4258@1498804 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/misc/src/java/org/apache/lucene/misc/SweetSpotSimilarity.java
@@ -158,7 +158,7 @@ public class SweetSpotSimilarity extends DefaultSimilarity {
    * @see #baselineTf
    */
   @Override
-  public float tf(int freq) {
+  public float tf(float freq) {
     return baselineTf(freq);
   }
